==> 3 - Protein compartment shifts.Rmd <==
---
title: "Compartment shifts"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load packages and data


```{r}
library(tidyverse)
library(openxlsx)

##Read in data
data <- read.xlsx("Input data/ProteinLevelRawRenamed.xlsx")  
LFQ_data <- data %>%
  filter(is.na(Reverse)) %>%
  filter(is.na(Potential.contaminant)) %>%
  select(Protein.IDs, LFQ_A1:LFQ_J7)

sample_key <- read.xlsx("Input data/SampleKey.xlsx")



LFQ_data_long <- LFQ_data %>%
  pivot_longer(!Protein.IDs, names_to = "SampleID", values_to = "LFQ")

LFQ_data_long$LFQ[LFQ_data_long$LFQ == 0] <- NA

LFQ_data_long <- merge(sample_key, LFQ_data_long, by = "SampleID")

```

##Express all LFQ values as proportion of total

```{r}


##Use all fractions, don't impute missing values
NA_summary <- LFQ_data_long %>%
  group_by(Protein.IDs) %>%
  summarize(n_NA = sum(is.na(LFQ))) 

NA_summary_filtered <- NA_summary %>%
  filter(n_NA <= 14)

ratios <- LFQ_data_long %>%
  filter(Protein.IDs %in% NA_summary_filtered$Protein.IDs) %>%
  mutate_all(~ifelse(is.na(.x), 0.0000000001, .x)) %>%
  select(Subject, Protein.IDs, Condition, Fraction, LFQ) %>%
  pivot_wider(names_from = Fraction, values_from = LFQ) %>%
  mutate(total_LFQ = `1` + `2` + `3`+ `4` + `5` + `6` + `7`) %>%
  pivot_longer(-c(Subject, Protein.IDs, Condition, total_LFQ), names_to = "Fraction", values_to = "LFQ") %>%
  mutate(ratio = LFQ/total_LFQ)


```


###Consistency of proportions

```{r}
ratio_CV <- ratios %>%
  group_by(Condition, Protein.IDs, Fraction) %>%
  summarize(mean_ratio = mean(ratio), sd_ratio = sd(ratio)) %>%
  mutate(CV = (sd_ratio/mean_ratio)*100)


ggplot(ratio_CV, aes(x = CV)) + geom_histogram(binwidth = 5) + facet_grid(rows = vars(Condition), cols = vars(Fraction)) + 
  geom_vline(xintercept = 20, color = "red") + xlab("Coefficient of variation") + ylab("Number of proteins")

## check ratio distributions per fraction and condition
ratios %>%
  ggplot(aes(x=ratio, colour=Condition)) +geom_density() +facet_wrap(~Fraction)

## how many obs per protein per fraction per group - calc means and stdevs
ratio_summary = ratios %>%
  filter(!grepl("CON_", Protein.IDs)) %>%
  filter(!grepl("REV_", Protein.IDs)) %>%
  group_by(Protein.IDs, Fraction) %>%
  summarise(n=n(),
            n_C=length(which(Condition=="C")),
            n_AD=length(which(Condition=="AD")),
            mean_C=mean(ratio[Condition=="C"]),
            stdev_C=sd(ratio[Condition=="C"]),
            mean_AD=mean(ratio[Condition=="AD"]),
            stdev_AD=sd(ratio[Condition=="AD"]))

```

##Calculate entropy for each set of proportions

```{r}
res_entropy = ratios %>%
  filter(!grepl("CON_", Protein.IDs)) %>%
  filter(!grepl("REV_", Protein.IDs)) %>%
  group_by(Protein.IDs,Subject,Condition) %>%
  summarise(n=n(),
            entropy=-sum(ratio * log2(ratio))/log2(n)) %>%
  ungroup() %>%
  group_by(Protein.IDs) %>%
  summarise(n=n(),
            n_C=length(which(Condition=="C")),
            n_AD=length(which(Condition=="AD")),
            mean_C=mean(entropy[Condition=="C"]),
            mean_AD=mean(entropy[Condition=="AD"]),
            diff=abs(mean_AD-mean_C),
            stdev_C=sd(entropy[Condition=="C"]),
            stdev_AD=sd(entropy[Condition=="AD"]),
            stat_t=t.test(entropy[Condition=="C"],
                          entropy[Condition=="AD"])$statistic,
            stat_w=wilcox.test(entropy[Condition=="C"],
                               entropy[Condition=="AD"])$statistic,
            pVal_t=t.test(entropy[Condition=="C"],
                          entropy[Condition=="AD"])$p.value,
            pVal_w=wilcox.test(entropy[Condition=="C"],
                               entropy[Condition=="AD"])$p.value) %>%
  arrange(desc(abs(stat_t)))


head(res_entropy)

```



```{r}


w_sig <- res_entropy %>%
  filter(pVal_w < 0.1)

# plot pValue distributions
res_entropy %>%
  ggplot(aes(x=pVal_t)) +geom_histogram() +
  theme(legend.position="none") +
  geom_vline(xintercept = 0.05)



```

##T-tests for entropy differences


```{r}
##
## Perform t-tests
##
res_tTest = ratios %>%
  filter(!grepl("CON_", Protein.IDs)) %>%
  filter(!grepl("REV_", Protein.IDs)) %>%
  group_by(Protein.IDs, Fraction) %>%
  summarise(n=n(),
            n_C=length(which(Condition=="C")),
            n_AD=length(which(Condition=="AD")),
            tStat=t.test(ratio[Condition=="C"], 
                        ratio[Condition=="AD"])$statistic,
            pVal=t.test(ratio[Condition=="C"], 
                        ratio[Condition=="AD"])$p.value) %>%
  mutate(padj = p.adjust(pVal, method = "BH")) %>%
  arrange(padj)

head(res_tTest)



# plot pValue distributions
res_tTest %>%
  ggplot(aes(x=pVal,fill=Fraction)) +geom_histogram() +
  facet_wrap(~Fraction, ncol=1) +theme(legend.position="none") +
  geom_vline(xintercept = 0.05)
```

##Interesting proteins


```{r}
# Look at the top hits
# res_tTest
results <- res_tTest %>%
  mutate(ttest_significant = ifelse(padj < 0.05, "Yes", "No")) 
  
entropy <- res_entropy %>%
  select(Protein.IDs, mean_C, mean_AD, diff) %>%
  rename(`Entropy Control` = mean_C) %>%
  rename(`Entropy AD` = mean_AD) %>%
  rename(`Entropy difference` = diff)

results <- results %>%
  left_join(entropy) %>%
  arrange(Protein.IDs)
  

write_csv(results, "Supp tables/TableS6_ttest_outcomes.csv")

interesting <- c("sp|P54750|PDE1A_HUMAN", "sp|P10636|TAU_HUMAN", "sp|Q96F85|CNRP1_HUMAN", "sp|P49841|GSK3B_HUMAN")

# plot from the original ratios
interesting_ratios <- ratios %>%
  filter(Protein.IDs %in% interesting) %>%
  group_by(Protein.IDs) %>%
  dplyr::mutate(prot = gsub("_HUMAN","",strsplit(Protein.IDs,"|",fixed=T)[[1]][3])) %>%
  ungroup() %>%
  ggplot(aes(x=Fraction,y=ratio,colour=Condition)) + ylab("proportion in fraction") +
  geom_hline(yintercept = 1/7, alpha=0.5) +
  geom_boxplot(alpha=0.1, width=0.5, position = position_dodge(0.5)) + geom_point(aes(alpha = 0.2), shape = 16, position = position_dodge(0.5)) + 
  facet_wrap(~prot, scales="fixed", ncol = 4) + theme_bw()


ggsave("Figures/Fig3B_ratios_interesting_proteins.pdf", plot = interesting_ratios, device = "pdf", width = 8.5, height = 4, units = "in", dpi = 300)

interesting_ratios

```


##LFQs of interesting proteins

```{r}


LFQ_interesting <- LFQ_data_long %>%
  filter(Protein.IDs %in% interesting) %>%
  separate(Protein.IDs, into = c(NA, "UNIPROT", "DESC"), sep = "\\|", extra = "drop") %>%
  ggplot(aes(x=as.factor(Fraction),y=log2(LFQ),colour=Condition)) +
    geom_boxplot(alpha=0.1, width=0.5, position = position_dodge(0.5)) + geom_point(aes(alpha = 0.2), shape = 16, position = position_dodge(0.5)) + 
    facet_wrap(~DESC, scales="free_y", ncol = 4) + theme_bw()

ggsave("Figures/Fig3A_LFQ_interesting_proteins.pdf", plot = LFQ_interesting, device = "pdf", width = 8.5, height = 4, units = "in", dpi = 300)

LFQ_interesting

```

##Summed values across fractions

```{r}

summed <- LFQ_data_long %>%
  filter(Protein.IDs %in% interesting) %>%
  separate(Protein.IDs, into = c(NA, "UNIPROT", "DESC"), sep = "\\|", extra = "drop") %>%
  group_by(Subject, DESC, Condition) %>%
  filter(!is.na(LFQ)) %>%
  summarise(total_LFQ = sum(LFQ)) %>%
  ggplot(aes(x=Condition, y=log2(total_LFQ),colour=Condition)) + geom_point() +
  geom_boxplot(alpha=0.1, width=0.5) +
  facet_wrap(~DESC, ncol = 4) + theme_bw()



ggsave("Figures/Fig3C_LFQ_summed.pdf", plot = summed, device = "pdf", width = 8.5, height = 3, units = "in", dpi = 300)


summed
```

##Consistency of distributin through fractions - RMSE

```{r}

##RMSE for consistency of distribution through fractions

head(ratios)


ratio_averages <- ratios %>%
  group_by(Condition, Protein.IDs, Fraction) %>%
  summarize(mean_ratio = mean(ratio)) %>%
  left_join(ratios, by = c("Condition", "Protein.IDs", "Fraction")) %>%
  mutate(diff_squared = (ratio - mean_ratio)^2) %>%
  group_by(Protein.IDs, Condition) %>%
  summarize(RMSE = sqrt(sum(diff_squared)/7))

ggplot(ratio_averages, aes(x = RMSE)) + geom_histogram(binwidth = 0.01) + facet_wrap(~Condition)

ggplot(ratio_averages, aes(x = RMSE, group = Condition)) + geom_density(aes(color = Condition)) 



##Stratify by confidence
gold <- read_csv("Input data/gold_markers.csv")
marker_confidence <- gold %>%
  mutate(Confidence = "High") %>%
  select(ProteinID, Confidence) %>%
  mutate(Condition = "AD") %>%
  select(ProteinID, Condition, Confidence)
marker_confidence_con <- gold %>%
  mutate(Confidence = "High") %>%
  select(ProteinID, Confidence) %>%
  mutate(Condition = "Control") %>%
  select(ProteinID, Condition, Confidence) %>%
  bind_rows(marker_confidence)


confidence <- read_csv("Supp tables/Table_S5_SVM_results.csv")
confidence <- confidence %>%
  select(Uniprot_ID, Confidence_control, Confidence_AD) %>%
  pivot_longer(-Uniprot_ID, names_to = "Condition", values_to = "Confidence") %>%
  rename(ProteinID = Uniprot_ID) %>%
  mutate(Condition = gsub("Confidence_control", "Control", Condition)) %>%
  mutate(Condition = gsub("Confidence_AD", "AD", Condition)) %>%
  bind_rows(marker_confidence_con)







RMSE_to_plot <- ratio_averages %>%
  separate(Protein.IDs, into = c(NA, "ProteinID", NA))  %>%
  mutate(Condition = gsub("C", "Control", Condition))
RMSE_to_plot <- merge(RMSE_to_plot, confidence, by = c("ProteinID", "Condition"))

confidence_col <- c("High" = "#2c7bb6", "Medium" = "#fdae61", "Low" = "#d7191c")

RMSE_to_plot <- RMSE_to_plot %>%
  filter(Confidence != "None")
RMSE_to_plot$Confidence <- factor(RMSE_to_plot$Confidence, levels = c("High", "Medium", "Low"))
levels(RMSE_to_plot$Confidence)

RMSE <- RMSE_to_plot %>%
  filter(!is.na(Condition)) %>%
  ggplot(aes(x = RMSE, group = Confidence)) + geom_density(aes(color = Confidence)) +
  scale_color_manual(values = confidence_col) + facet_wrap(~Condition, ncol = 1)

RMSE_to_plot %>%
  filter(Confidence != "None") %>%
  filter(!is.na(Condition)) %>%
  ggplot(aes(x = RMSE)) + geom_histogram(binwidth = 0.01) + facet_grid(rows = vars(Confidence), cols = vars(Condition))

ggsave("Figures/FigS5C_RMSE.pdf", plot = RMSE, device = "pdf", width = 8.5, height = 4, units = "in", dpi = 300)

RMSE
```
